requirements.base.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
kafka-python==2.0.2
neo4j==5.19.0
sentence-transformers==2.7.0
scikit-learn==1.5.0
openai==1.23.2
google-ads==22.0.0
facebook-business==18.0.8
pydantic-settings==2.2.1
python-dotenv==1.0.1
black==24.4.0
ruff==0.4.4
pytest==8.1.1
requests==2.31.0
